- Code for culling/displaying WorldGrid 	- Doesn't currently work so debug ASAP. - Some debug code in WorldGrid for drawing a moving a line. - Fix in Debug as was trying to use a uint8 for values up to 17*17. - flVec2 is now templated.

+++ source/Debug.cpp
@@ -262,7 +262,7 @@ void Debug::UnpropagateAdjacent()
 void Debug::ResetWorldGrid()
 {
 	uint32_t centerIndex = m_debugWorldDim % 2 ? m_debugWorldDim / 2 + 1 : m_debugWorldDim / 2;
-	for (uint8_t i = 0; i < m_debugWorldDim*m_debugWorldDim; i++)
+	for (uint16_t i = 0; i < m_debugWorldDim*m_debugWorldDim; i++)
 	{
 		if (i != centerIndex)
 		{
